+++ savu/plugins/filters/denoising/base_median_filter.py
@@ -23,20 +23,27 @@
 
 import numpy as np
 from savu.plugins.plugin import Plugin
+from savu.plugins.driver.cpu_plugin import CpuPlugin
+from savu.core.iterate_plugin_group_utils import enable_iterative_loop, \
+    check_if_end_plugin_in_iterate_group, setup_extra_plugin_data_padding
 
-class BaseMedianFilter(Plugin):
+class BaseMedianFilter(Plugin, CpuPlugin):
 
     def __init__(self, name='BaseMedianFilter'):
         super(BaseMedianFilter, self).__init__(name)
         self.frame_limit = 1
 
+    @enable_iterative_loop
     def setup(self):
         in_dataset, out_dataset = self.get_datasets()
         out_dataset[0].create_dataset(in_dataset[0])
         in_pData, out_pData = self.get_plugin_datasets()
-        in_pData[0].plugin_data_setup(self.parameters['pattern'], self.get_max_frames())
-        out_pData[0].plugin_data_setup(self.parameters['pattern'], self.get_max_frames())
+        in_pData[0].plugin_data_setup(self.parameters['pattern'],
+                                        self.get_max_frames())
+        out_pData[0].plugin_data_setup(self.parameters['pattern'],
+                                        self.get_max_frames())
 
+    @setup_extra_plugin_data_padding
     def set_filter_padding(self, in_data, out_data):
         # kernel size must be odd
         ksize = self.parameters['kernel_size']
@@ -52,3 +59,17 @@ class BaseMedianFilter(Plugin):
     def get_max_frames(self):
         """ Setting nFrames to multiple with an upper limit of 4 frames. """
         return ['multiple', self.frame_limit]
+
+    # total number of output datasets
+    def nOutput_datasets(self):
+        if check_if_end_plugin_in_iterate_group(self.exp):
+            return 2
+        else:
+            return 1
+
+    # total number of output datasets that are clones
+    def nClone_datasets(self):
+        if check_if_end_plugin_in_iterate_group(self.exp):
+            return 1
+        else:
+            return 0