+++ savu/plugins/filters/denoising/ccpi_denoising_gpu_tools.py
@@ -118,11 +118,11 @@ efficient 2D/3D denoising.
              default: 0.0
 
         pattern:
-             visibility: advanced
+             visibility: intermediate
              dtype: str
              description: Pattern to apply this to.
+             options: [SINOGRAM, PROJECTION, VOLUME_XZ, VOLUME_XY]
              default: 'VOLUME_XZ'
-
         """